iOS更新9.5

+++ Demo/TXLiteAVDemo/TRTCDemo/UI/SettingCells/TRTCSettingsSliderCell.m
@@ -1,20 +1,21 @@
 /*
-* Module:   TRTCSettingsSliderCell
-*
-* Function: 配置列表Cell，右侧是一个Slider
-*
-*/
+ * Module:   TRTCSettingsSliderCell
+ *
+ * Function: 配置列表Cell，右侧是一个Slider
+ *
+ */
 
 #import "TRTCSettingsSliderCell.h"
+
 #import "Masonry.h"
-#import "UISlider+TRTC.h"
 #import "UILabel+TRTC.h"
+#import "UISlider+TRTC.h"
 
 @interface TRTCSettingsSliderCell ()
 
-@property (strong, nonatomic) UISlider *slider;
-@property (strong, nonatomic) UILabel *valueLabel;
-@property (strong, nonatomic) NSNumberFormatter *numFormatter;
+@property(strong, nonatomic) UISlider *         slider;
+@property(strong, nonatomic) UILabel *          valueLabel;
+@property(strong, nonatomic) NSNumberFormatter *numFormatter;
 
 @end
 
@@ -22,13 +23,13 @@
 
 - (void)setupUI {
     [super setupUI];
-    
+
     self.slider = [UISlider trtc_slider];
     [self.slider addTarget:self action:@selector(onSliderValueChange:forEvent:) forControlEvents:UIControlEventValueChanged];
     [self.contentView addSubview:self.slider];
 
-    self.valueLabel = [UILabel trtc_contentLabel];
-    self.valueLabel.textAlignment = NSTextAlignmentRight;
+    self.valueLabel                           = [UILabel trtc_contentLabel];
+    self.valueLabel.textAlignment             = NSTextAlignmentRight;
     self.valueLabel.adjustsFontSizeToFitWidth = YES;
     [self.contentView addSubview:self.valueLabel];
 
@@ -37,19 +38,19 @@
         make.trailing.equalTo(self.valueLabel.mas_leading).offset(-5);
         make.width.mas_equalTo(180);
     }];
-    
+
     [self.valueLabel mas_makeConstraints:^(MASConstraintMaker *make) {
         make.centerY.equalTo(self.contentView);
         make.trailing.equalTo(self.contentView).offset(-18);
         make.width.mas_equalTo(36);
     }];
-    
+
     [self.titleLabel mas_remakeConstraints:^(MASConstraintMaker *make) {
         make.left.equalTo(self.contentView).offset(9);
         make.top.equalTo(self.contentView).offset(15);
     }];
-    
-    self.numFormatter = [[NSNumberFormatter alloc] init];
+
+    self.numFormatter                       = [[NSNumberFormatter alloc] init];
     self.numFormatter.minimumFractionDigits = 0;
     self.numFormatter.maximumFractionDigits = 1;
 }
@@ -69,29 +70,29 @@
         if (0 == sliderItem.step) sliderItem.step = 1.f;
         self.slider.minimumValue = sliderItem.minValue / sliderItem.step;
         self.slider.maximumValue = sliderItem.maxValue / sliderItem.step;
-        self.slider.value = sliderItem.sliderValue / sliderItem.step;
-        self.slider.continuous = sliderItem.continuous;
-        self.valueLabel.text = [self.numFormatter stringFromNumber:@(sliderItem.sliderValue)];
+        self.slider.value        = sliderItem.sliderValue / sliderItem.step;
+        self.slider.continuous   = sliderItem.continuous;
+        self.valueLabel.text     = [self.numFormatter stringFromNumber:@(sliderItem.sliderValue)];
     }
-    
+
     [item addObserver:self forKeyPath:@"sliderValue" options:NSKeyValueObservingOptionNew context:nil];
 }
 
-- (void)observeValueForKeyPath:(NSString *)keyPath ofObject:(id)object change:(NSDictionary<NSKeyValueChangeKey,id> *)change context:(void *)context {
+- (void)observeValueForKeyPath:(NSString *)keyPath ofObject:(id)object change:(NSDictionary<NSKeyValueChangeKey, id> *)change context:(void *)context {
     if ([keyPath isEqualToString:@"sliderValue"]) {
-        TRTCSettingsSliderItem *item = (TRTCSettingsSliderItem *) self.item;
+        TRTCSettingsSliderItem *item = (TRTCSettingsSliderItem *)self.item;
         if (0 == item.step) item.step = 1.f;
-        self.slider.value = item.sliderValue / item.step;
+        self.slider.value    = item.sliderValue / item.step;
         self.valueLabel.text = [self.numFormatter stringFromNumber:@(item.sliderValue)];
     }
 }
 
-- (void)onSliderValueChange:(UISlider *)slider forEvent:(UIEvent*)event {
+- (void)onSliderValueChange:(UISlider *)slider forEvent:(UIEvent *)event {
     UITouch *touchEvent = [[event allTouches] anyObject];
-    
+
     TRTCSettingsSliderItem *sliderItem = (TRTCSettingsSliderItem *)self.item;
-    int value = slider.value * sliderItem.step;
-    self.valueLabel.text = [self.numFormatter stringFromNumber:@(value)];
+    float                   value      = slider.value * sliderItem.step;
+    self.valueLabel.text               = [self.numFormatter stringFromNumber:@(value)];
 
     switch (touchEvent.phase) {
         case UITouchPhaseEnded:
@@ -105,24 +106,17 @@
 
 @end
 
-
 @implementation TRTCSettingsSliderItem
 
-- (instancetype)initWithTitle:(NSString *)title
-                        value:(float)value
-                          min:(float)min
-                          max:(float)max
-                         step:(float)step
-                   continuous:(BOOL)continuous
-                       action:(void (^)(float))action {
+- (instancetype)initWithTitle:(NSString *)title value:(float)value min:(float)min max:(float)max step:(float)step continuous:(BOOL)continuous action:(void (^)(float))action {
     if (self = [super init]) {
-        self.title = title;
+        self.title   = title;
         _sliderValue = value;
-        _minValue = min;
-        _maxValue = max;
-        _step = step == 0 ? 1 : step;
-        _continuous = continuous;
-        _action = action;
+        _minValue    = min;
+        _maxValue    = max;
+        _step        = step == 0 ? 1 : step;
+        _continuous  = continuous;
+        _action      = action;
     }
     return self;
 }
@@ -136,4 +130,3 @@
 }
 
 @end
-
